rag_app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymongo
python-dotenv
faiss-cpu
openai
psycopg2-binary
